% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/animation.R
\name{animation}
\alias{animation}
\title{Display Animation of Control Strength of each time-point or segment}
\usage{
animation(obj, segment = FALSE, filename = "index.html", width = 1000,
  height = 1000)
}
\arguments{
\item{obj}{an object of class tacs}

\item{segment}{logical if TRUE, plot control strength of each segment, FALSE plot each gene}

\item{filename}{filename of html that includes the animation.}

\item{width}{width of animation}

\item{height}{height of animation}
}
\description{
Display Animation of Control Strength of each time-point or segment
}
\examples{
data(dream4)
res <- tacs(dream4$net,dream4$exp,"aic-g",TRUE,dream4$segment)
animation(res)
}

