% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ic.R
\name{calc_ic}
\alias{calc_ic}
\title{Calculate information criterion of each node of the network without edge,
and ic of every innode for every time point}
\usage{
calc_ic(exp_data, info)
}
\arguments{
\item{exp_data}{experimental data on nodes in the network: data frame}

\item{info}{input parameters of tacs execution except experimental data}
}
\value{
Scores of node controll strength and supplementary information
}
\description{
Calculate information criterion of each node of the network without edge,
and ic of every innode for every time point
}

