% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency_check.R
\name{consistency_check}
\alias{consistency_check}
\title{Check the consistency on input parameters}
\usage{
consistency_check(exp_data, ic_type, segment)
}
\arguments{
\item{exp_data}{experimental data on nodes in the network: data frame}

\item{ic_type}{information criterion type: string "aic"(=default) or "bayes"}

\item{segment}{This paramer assigns the segments of exp_data. 
It is used when exp_data is composed of plural time-course experiment and the edges are drown under markov property}
}
\value{
Boolean
}
\description{
Check the consistency on input parameters
}

