% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_edge_weight.R
\name{cutoff_edge_weight}
\alias{cutoff_edge_weight}
\title{set edges weights to zero under n*standard deviations}
\usage{
cutoff_edge_weight(elist, sd, n)
}
\arguments{
\item{elist}{edge list with weights. It have 3 columns [from, to, weight].}

\item{sd}{standared deviations of edges}

\item{n}{deviation}
}
\value{
a vector of new edge weights
}
\description{
set edges weights to zero under n*standard deviations
}

