% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dream4.R
\docType{data}
\name{dream4}
\alias{dream4}
\title{time-course gene expression profile with 10 genes from dream4.}
\format{a list. $exp: gene expression profile, $net: gene regulatory network, $segment: 21 time points x 5 experiments, $comment:comment to get acyclic graph}
\source{
DREAM4 - In Silico Network Challenge
}
\usage{
data(dream4)
}
\description{
This data set includes one time-course gene expression profile
used in Dream4 in silico network challenge held in 2009.
}
\author{
Marbach D, Prill RJ, Schaffter T, Mattiussi C, Floreano D, and Stolovitzky G. Revealing strengths and weaknesses of methods for gene network inference. PNAS, 107(14):6286-6291, 2010.
Marbach D, Schaffter T, Mattiussi C, and Floreano D. Generating Realistic in silico Gene Networks for Performance Assessment of Reverse Engineering Methods. Journal of Computational Biology, 16(2) pp. 229-239, 2009.
Prill RJ, Marbach D, Saez-Rodriguez J, Sorger PK, Alexopoulos LG, Xue X, Clarke ND, Altan-Bonnet G, and Stolovitzky G. Towards a rigorous assessment of systems biology models: the DREAM3 challenges. PLoS ONE, 5(2):e9202, 2010.
}

