% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.tacs.R
\name{plot.tacs}
\alias{plot.tacs}
\title{Plot function for tacs class}
\usage{
\method{plot}{tacs}(obj, segment = FALSE, max_node = 0, ...)
}
\arguments{
\item{obj}{an object of class tacs}

\item{segment}{logical if TRUE, plot control strength of each segment, FALSE plot each gene}

\item{max_node}{maximum number of nodes to display}
}
\description{
Plot function for tacs class
}
\examples{
data(dream4)
res <- tacs(dream4$net,dream4$exp,"aic-g",TRUE,dream4$segment)
plot(res)
}

