% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tacs.R
\name{print.tacs}
\alias{print.tacs}
\title{print function for tacs class}
\usage{
\method{print}{tacs}(obj, ...)
}
\arguments{
\item{obj}{an object of class tacs}
}
\description{
print function for tacs class
}
\examples{
data(dream4)
res <- tacs(dream4$net,dream4$exp,"aic-g",TRUE,dream4$segment)
print(res)
}

