% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select.node.R
\name{select.nodes}
\alias{select.nodes}
\title{Select nodes by name list}
\usage{
select.nodes(obj, nodes, delete = FALSE)
}
\arguments{
\item{obj}{an object of class tacs}

\item{nodes}{nodes names to be selected.}

\item{delete}{if TRUE designated nodes are deleted}
}
\description{
Select nodes by name list
}
\examples{
data(dream4)
obj <- tacs(dream4$net,dream4$exp,"aic-g",TRUE,dream4$segment)
nodes <- c("mtlR","gadW","marR")
select1 <- select.nodes(obj,nodes)
select2 <- select.nodes(obj,nodes,delete=TRUE)
}

