% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select.segments.R
\name{select.segments}
\alias{select.segments}
\title{Select segment by name}
\usage{
select.segments(obj, segs, delete = FALSE)
}
\arguments{
\item{obj}{an object of class tacs}

\item{segs}{segment name list to be selected.}

\item{delete}{if TRUE designated segments are deleted}
}
\description{
Select segment by name
}
\examples{
data(dream4)
obj <- tacs(dream4$net,dream4$exp,"aic-g",TRUE,dream4$segment)
segs <- c("seg1","seg4")
select1 <- select.segments(obj,segs)
select2 <- select.segments(obj,segs,delete=TRUE)
}

