% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/set_edge_weight.R
\name{set_edge_weight}
\alias{set_edge_weight}
\title{generate edgeweight from a row of dataframe}
\usage{
set_edge_weight(elist, node_weight)
}
\arguments{
\item{elist}{edge list}

\item{node_weight}{node weight}
}
\value{
a vector of edge weights
}
\description{
generate edgeweight from a row of dataframe
}

