% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/set_layout.R
\name{set_layout}
\alias{set_layout}
\title{returns layout of graph nodes}
\usage{
set_layout(obj)
}
\arguments{
\item{obj}{an object of class tacs}
}
\value{
matrix with two columns. each row shows a coordinates of a node.
}
\description{
returns layout of graph nodes
}

