% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_markov_exp.R
\name{set_lower_exp}
\alias{set_lower_exp}
\title{Convert data set for simple markov property, set data for down stream node}
\usage{
set_lower_exp(exp_data, segment)
}
\arguments{
\item{exp_data}{experimental data on nodes in the network: data frame}

\item{segment}{This paramer assigns the segments of exp_data. 
It is used when exp_data is composed of plural time-course experiment and the edges are drown under markov property}
}
\value{
data set without segment start
}
\description{
Convert data set for simple markov property, set data for down stream node
}

