% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/set_row_name.R
\name{set_row_name}
\alias{set_row_name}
\title{Generate list of row names of data frame, which to store ic of each node
When is_markov is TRUE, the simple markov model is adopted.}
\usage{
set_row_name(sample_name, is_markov, segment)
}
\arguments{
\item{sample_name}{: rowname of data_frame}

\item{is_markov}{edges are drown under markov property :boolean TRUE or FALSE(=default)}

\item{segment}{This paramer assigns the segments of exp_data.
It is used when exp_data is composed of plural time-course experiment and the edges are drown under markov property}
}
\value{
row name (sample name) of data frame to store ic of each node
}
\description{
Generate list of row names of data frame, which to store ic of each node
When is_markov is TRUE, the simple markov model is adopted.
}

