% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_segname_col.R
\name{set_segname_col}
\alias{set_segname_col}
\title{Generate list of segment name to categorize}
\usage{
set_segname_col(is_markov, segment, segment_name)
}
\arguments{
\item{is_markov}{edges are drown under markov property :boolean TRUE or FALSE(=default)}

\item{segment}{This paramer assigns the segments of exp_data. 
It is used when exp_data is composed of plural time-course experiment and the edges are drown under markov property}

\item{segment_name}{name list of segments. default (seg1,seg2,...)}
}
\value{
list of segment name
}
\description{
Generate list of segment name to categorize
}

