% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary.tacs.R
\name{summary.tacs}
\alias{summary.tacs}
\title{summary function for tacs class}
\usage{
\method{summary}{tacs}(obj, ...)
}
\arguments{
\item{obj}{an object of class tacs}
}
\description{
summary function for tacs class
}
\examples{
data(dream4)
res <- tacs(dream4$net,dream4$exp,"aic-g",TRUE,dream4$segment)
summary(res)
}

