% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tacs.R
\docType{package}
\name{tacs}
\alias{tacs}
\alias{tacs-package}
\title{Calculate Control Strength toward every nodes on directed graph based on information criterion}
\usage{
tacs(network, tc_data, ic_type = "aic", is_markov = FALSE)
}
\arguments{
\item{network}{directed graph: format used in library bnlearn is used.}

\item{tc_data}{time course data on nodes in the network: data frame}

\item{ic_type}{information criterion type: string "aic"(=default) or "bayes"}

\item{is_markov}{edges are drown under markov property :boolean TRUE or FALSE(=default)}
}
\value{
strength Control strength toward each node for every time point: data frame
}
\description{
Calculate Control Strength toward every nodes on directed graph based on information criterion

Calculate Control Strength toward every nodes on directed graph based on information criterion
}

