% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vsCor.R
\name{vsCor}
\alias{vsCor}
\title{Calc coeffcient between Controll strength of nodes and experimental data}
\usage{
vsCor(obj, exp_data)
}
\arguments{
\item{obj}{an object of class tacs}

\item{exp_data}{experimental data use to calculate tacs: data frame}
}
\description{
Calc coeffcient between Controll strength of nodes and experimental data
}
\examples{
data(dream4)
res <- tacs(dream4$net,dream4$exp,"aic-g",TRUE,dream4$segment)
vsCor(res,dream4$exp)
}

