% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_ic.R
\name{calc_ic}
\alias{calc_ic}
\title{Calculate information criterion of each node of the network without edge,
and ic of every innode for every time point}
\usage{
calc_ic(network, exp_data, ic_type, is_markov, segment)
}
\arguments{
\item{network}{directed graph: format used in library bnlearn is used.}

\item{exp_data}{experimental data on nodes in the network: data frame}

\item{ic_type}{information criterion type: string "aic"(=default) or "bayes"}

\item{is_markov}{edges are drown under markov property :boolean TRUE or FALSE(=default)}

\item{segment}{This paramer assigns the segments of exp_data.
It is used when exp_data is composed of plural time-course experiment and the edges are drown under markov property}
}
\value{
two dataframes that include information criterion of nodes in the network without edge and
and ic of every innode for every time point
}
\description{
Calculate information criterion of each node of the network without edge,
and ic of every innode for every time point
}

