% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tacs.R
\docType{package}
\name{tacs}
\alias{tacs}
\alias{tacs-package}
\title{Calculate Control Strength toward every nodes on directed graph based on information criterion}
\usage{
tacs(network, exp_data, ic_type = "aic-g", is_markov = FALSE,
  segment = NA, segment_name = NA)
}
\arguments{
\item{network}{directed graph: format used in library bnlearn is used.}

\item{exp_data}{experimental data on nodes in the network: data frame}

\item{ic_type}{information criterion type: string "aic"(=default) or "bayes"}

\item{is_markov}{edges are drown under markov property :boolean TRUE or FALSE(=default)}

\item{segment}{This paramer assigns the segments of exp_data.
It is used when exp_data is composed of plural time-course experiment and the edges are drown under markov property}

\item{segment_name}{name list of segments. default (seg1,seg2,...)}
}
\value{
strength Control strength toward each node for every time point: data frame
}
\description{
Calculate Control Strength toward every nodes on directed graph based on information criterion

Calculate Control Strength toward every nodes on directed graph based on information criterion
}
\examples{
data(diauxie)#an example in GIW2015 paper
res <- tacs(diauxie$net,diauxie$exp,"aic-g",is_markov=TRUE,segment=diauxie$segment,segment_name=diauxie$segment_name)
gene_selected <- apply(res$score,2,max) + res$ave_InfoGain >0
score <- res$score[1:16,gene_selected] # 1:16: Wild Type.
score[score<0] <- 0  #to show the results in barplot, remove negatives
barplot(t(as.matrix(score)),width=nrow(score),beside=TRUE,
col=topo.colors(ncol(score)),legend.text=names(score))
WT <-  select.segments(res,"WT") #Plot score of each gene
WT_genes <- select.nodes(WT,names(gene_selected[gene_selected]))
plot(WT_genes)
data(dream4) #Anther example
tacs(dream4$net,dream4$exp,"aic-g",TRUE,dream4$segment,NA) -> res
}

